Added IHttpClientFactory and session support to Program.cs for admin login

+++ YurttaYe.Infrastructure/Persistence/Repositories/MenuItemRepository.cs
@@ -1,8 +1,10 @@
 // src/YurttaYe.Infrastructure/Persistence/Repositories/MenuItemRepository.cs
+using System.Collections.Generic;
+using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
-using YurttaYe.Application.Abstractions;
 using YurttaYe.Application.Abstractions.Repositories;
 using YurttaYe.Domain.Entities;
+using YurttaYe.Infrastructure.Persistence;
 
 namespace YurttaYe.Infrastructure.Persistence.Repositories
 {
@@ -22,26 +24,25 @@ namespace YurttaYe.Infrastructure.Persistence.Repositories
 
         public async Task<MenuItem> GetByIdAsync(int id)
         {
-            return await _context.MenuItems.FindAsync(id) ?? throw new Exception("Menu item not found");
+            return await _context.MenuItems.FindAsync(id);
         }
 
         public async Task AddAsync(MenuItem menuItem)
         {
             await _context.MenuItems.AddAsync(menuItem);
-            await _context.SaveChangesAsync();
         }
 
         public async Task UpdateAsync(MenuItem menuItem)
         {
             _context.MenuItems.Update(menuItem);
-            await _context.SaveChangesAsync();
+            await Task.CompletedTask;
         }
 
         public async Task DeleteAsync(int id)
         {
             var menuItem = await GetByIdAsync(id);
-            _context.MenuItems.Remove(menuItem);
-            await _context.SaveChangesAsync();
+            if (menuItem != null)
+                _context.MenuItems.Remove(menuItem);
         }
     }
 }
